refactor(html): separate a part of routing
