+++ void-base/src/main/kotlin/io/void/router/util/RequestHandler.kt
@@ -1,12 +1,9 @@
 package io.void.router.util
 
-import io.void.api.KtsPage
 import io.void.cache.Cache
 import io.void.clienthandler.ClientHandler
 import io.void.dto.http.RequestDTO
 import io.void.dto.http.ResponseDTO
-import io.void.dto.http.buildResponse
-import io.void.dto.http.headers
 import io.void.html.page.Page
 import io.void.html.page.dynamic.DynamicPage
 import io.void.html.page.dynamic.Path
